+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Set av1_vaapi default `--max-crf` to 255.
 * Fix sample-encode printing output to non-terminals.
 * Omit "Encode with: ..." stderr hint for non-terminals.
-* Support logging enable by setting env var `RUST_LOG`. E.g:
+* Support logging enabled when stderr is not a terminal or by setting env var `RUST_LOG`. E.g:
   - `RUST_LOG=ab_av1=info` "info" level logs various progress results like sample encode info
   - `RUST_LOG=ab_av1=debug` "debug" level logs include ffmpeg calls